+++ src/App.js
@@ -2,6 +2,7 @@ import React from "react";
 import "./App.css";
 import { useData } from "./utilities/firebase.js";
 import GroceryList from "./components/GroceryList.js";
+import UserGroceryList from "./components/UserGroceryList.js";
 import ButtonAppBar from "./components/AppBar.js";
 import AddNewItem from "./components/AddNewItem";
 
@@ -20,6 +21,10 @@ const App = () => {
         <GroceryList items={groceryList.items} users={groceryList.users} />
         <div>
           <AddNewItem />
+        </div>
+        My Own List
+        <div className="grocery-list">
+        <UserGroceryList items={groceryList.items} users={groceryList.users} />
         </div>
       </div>
     </div>
